+++ templates/page-profile-viewer.php
@@ -1,0 +1,42 @@
+<!doctype html>
+<html <?php language_attributes(); ?>
+<head>
+	<meta charset="<?php bloginfo( 'charset' ); ?>" />
+	<meta name="viewport" content="width=device-width, initial-scale=1" />
+	<?php wp_head(); ?>
+</head>
+
+<body <?php body_class(); ?>>
+	<?php wp_body_open(); ?>
+	<div id="page" class="site">
+		<div id="content" class="site-content">
+			<div id="primary" class="content-area">
+				<main id="main" class="site-main" role="main">
+					<?php
+					/* Start the Loop */
+					while ( have_posts() ) :
+						the_post();
+						?>
+						<article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
+
+							<header class="entry-header alignwide">
+								<?php the_title( '<h1 class="entry-title">', '</h1>' ); ?>
+							</header><!-- .entry-header -->
+
+							<div class="entry-content">
+								<?php the_content(); ?>
+							</div><!-- .entry-content -->
+
+						</article><!-- #post-<?php the_ID(); ?> -->
+						<?php
+					endwhile; // End of the loop.
+					?>
+				</div><!-- #main -->
+			</div><!-- #primary -->
+		</div><!-- #content -->
+	</div><!-- #page -->
+
+	<?php wp_footer(); ?>
+
+</body>
+</html>